fixed poc memory env (float64 -> float32)

+++ environments/poc_memory_env.py
@@ -54,7 +54,7 @@ class PocMemoryEnv():
         goals = np.asarray([-1.0, 1.0])
         # Determine the goal
         self._goals = goals[np.random.permutation(2)]
-        obs = np.asarray([self._goals[0], self._position, self._goals[1]])
+        obs = np.asarray([self._goals[0], self._position, self._goals[1]], dtype=np.float32)
         return obs
 
     @property
@@ -96,7 +96,7 @@ class PocMemoryEnv():
             self._position += self._step_size * (1 - self.freeze) if action == 1 else -self._step_size * (1 - self.freeze)
             self._position = np.round(self._position, 2)
 
-            obs = np.asarray([self._goals[0], self._position, self._goals[1]])
+            obs = np.asarray([self._goals[0], self._position, self._goals[1]], dtype=np.float32)
 
             if self.freeze: # Check if agent is allowed to move
                 self._step_count += 1
@@ -106,7 +106,7 @@ class PocMemoryEnv():
         else:
             self._position += self._step_size if action == 1 else -self._step_size
             self._position = np.round(self._position, 2)
-            obs = np.asarray([0.0, self._position, 0.0]) # mask out goal information
+            obs = np.asarray([0.0, self._position, 0.0], dtype=np.float32) # mask out goal information
 
         # Determine the reward function and episode termination
         if self._position == -1.0:
